+++ app/g/[groupSlug]/market/page.tsx
@@ -3,7 +3,7 @@
 import { redirect } from 'next/navigation';
 
 function Page() {
-  return redirect(`market/offer`);
+  return redirect(`market/sell`);
 }
 
 export default Page;
